requirements.txt
qrcode
pillow
pyinstaller